requirements.txt
django
django-cors-headers
django-filter
django-unfold
djangorestframework
djangorestframework-simplejwt
drf-redesign
drf-spectacular
gunicorn
loguru
markdown
psycopg2-binary
pydantic
pylint
pytest
pytest-django
python-dotenv
whitenoise
